% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LegalEntry.R
\name{legalentry}
\alias{legalentry}
\title{Legal Enrty}
\usage{
legalentry(train, test)
}
\arguments{
\item{train}{a vector of training strings}

\item{test}{a vector of test strings}
}
\value{
a vector of legal entries values for each test string
}
\description{
Computes whether the test string contains a legal entry (0 = illegal, 1 = legal)
}

