% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LevDist.R
\name{levdist}
\alias{levdist}
\title{Levenshtein (edit) Distance}
\usage{
levdist(training, test, measure)
}
\arguments{
\item{training}{a vector of training strings}

\item{test}{a vector of test strings}

\item{measure}{the measure to be used (mean, median, min, max)}
}
\value{
a vector containing the Levenshtein (edit) Distance of each test string
}
\description{
Computes Levenshtein (Edit) Distance
}

