% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern.R
\name{pattern}
\alias{pattern}
\title{Pattern Matching}
\usage{
pattern(train, test)
}
\arguments{
\item{train}{a vector of training strings}

\item{test}{a vector of test strings}
}
\value{
a vector containing the number of pattern matches that occur between the test item and all training items
}
\description{
Computes the number of times the pattern of the test string (regardless of surface structure) appears in training strings
}
