% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redundancy.R
\name{redundancy}
\alias{redundancy}
\title{Redundancy}
\usage{
redundancy(test)
}
\arguments{
\item{test}{a vector of test strings}
}
\value{
a vector containing the first-order redundancy of each test string
}
\description{
Computes first-order redundancy of each letter string
}
