% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringLength.R
\name{strlength}
\alias{strlength}
\title{String Length}
\usage{
strlength(test)
}
\arguments{
\item{test}{a vector of test strings}
}
\value{
a vector containing the length of each test string
}
\description{
Computes the length of each letter string
}

