% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheWorks.R
\name{theworks}
\alias{theworks}
\title{The Works}
\usage{
theworks(train, test)
}
\arguments{
\item{train}{a vector of training strings}

\item{test}{a vector of test strings}
}
\value{
a matrix containing string length, legal entry, six measures of novelty, six measures of ACS (Associative Chunk Strength), four measures of Levenshtein (edit) distance, first-order redundancy, and analogical similarity
}
\description{
Computes all item-level statistics
}
